build: Move to new zip packaging method
The build script now can package a zip without going through the whole process of building again.
The zip is now signed, as well as being able to use a new method to patch the zImage of the boot.img that the device has in flash.

This includes SignAPK, so credit to the maintainers of that tool.